+++ src/Network/Network/ChangeLog.md
@@ -19,7 +19,7 @@
 --->
 
 ## Upcoming Release
-* Added support to associate a DDoS custom policy (DCP) with an instance-level Public IP address.
+* Added support to associate a DDoS custom policy (DCP) with a supported Public IP address attachment.
     - Added the `-DdosCustomPolicyId` parameter to `Set-AzPublicIpAddress`.
     - Added the `-RemoveDdosCustomPolicy` switch to remove an existing association.
 * Added new cmdlets for ConnectionPolicy management under VirtualHub